requirements.txt
langchain
langchain-openai
langchain-community
sqlalchemy
python-dotenv
pyodbc